requirements.txt
Django
django_grappelli
django-crispy-forms
django-bootstrap-v5